fix(round6): resolve 4 bug-hunt defects; #4 was a false positive
Full suite green at 829 tests with zero regressions.

#1 [medium] Generated sub-service dispatch cleanup now wraps the orphaned
   instance's DisposeAsync/Dispose in a best-effort nested try/catch before
   rethrowing, so a faulting disposer can no longer replace the original
   registration failure on the wire and in diagnostics. (Dispatcher snapshot
   re-accepted accordingly.)

#2 [medium] IdentifierHelpers.EscapeIdentifier also consults
   GetContextualKeywordKind, so methods/parameters named with a contextual
   keyword (async/await/record/...) are @-escaped and the generated code
   compiles.

#3 [low] PooledBufferWriter.Advance widens its bounds check to 64-bit,
   symmetric with the R5#4 EnsureCapacity fix, so a large count can no longer
   overflow the guard and silently corrupt _written.

#6 [low] ProxyGenerator unescapes the method name in the NotSupportedException
   message (UnescapeIdentifier), so it reads 'event' rather than '@event'.

#4 FALSE POSITIVE (caught by the full-suite green gate): treating a truncated
   length prefix as a protocol error contradicts the deliberate, consistently
   tested policy (StreamConnectionCoverageTests.ReceiveAsync_ReturnsEmpty_
   When{PrefixTruncated,BodyTruncated,OnImmediateEof}) that any incomplete
   frame at EOF is an orderly disconnect (Payload.Empty). Fixing only the
   prefix case would also be inconsistent with the body case. Reverted the
   source change and removed its RED test.

+++ src/ShaRPC.Core/Buffers/PooledBufferWriter.cs
@@ -45,7 +45,9 @@ public sealed class PooledBufferWriter : IBufferWriter<byte>, IDisposable
         }
 
         var buffer = _buffer ?? throw new ObjectDisposedException(nameof(PooledBufferWriter));
-        if (_written + count > buffer.Length)
+        // Widen to 64-bit: _written + count in 32-bit signed arithmetic can overflow to a negative value,
+        // making this guard pass and silently corrupting _written (symmetric with the EnsureCapacity fix).
+        if ((long)_written + count > buffer.Length)
         {
             throw new InvalidOperationException("Cannot advance past the end of the buffer.");
         }
